+++ src/darkseed/dns/__init__.py
@@ -1,13 +1,12 @@
 """Module for DNS-related functionality."""
 
-from .null_record import NullRecord, NullRecordCodec
+from .aaaa_codec import AAAACodec
 from .regular_records import RegularRecords
 from .server import DNSConstants, DNSServer
 
 __all__ = [
+    "AAAACodec",
     "DNSConstants",
     "DNSServer",
     "RegularRecords",
-    "NullRecord",
-    "NullRecordCodec",
 ]